Removed unneeded services and modified existing to better fit new controllers
